bugfix+refactor+feat: ultimate test mode
was too confusing having multiple test modes.. went with: Repository Factory Disposed as the defacto mode. All data will be aborted/deleted once the factory has been disposed. Several changes to accommodate this behavior.

also allow some new session options to be specified, such as read-only and secondary-preferred for snapshots, and read timeout for transactions

refactor so that there is _always_ a mongo session, it just behaves differently in read-only/write mode

refactor commands into an object, and put dependencies in the mongo session so that they aren't passed around over and over and over